pyproject.toml
[project]
name = "agentic-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aif360>=0.5.0",
    "fairlearn>=0.7.0",
    "google-cloud-bigquery>=3.10.0",
    "google-cloud-bigquery-storage>=2.19.0",
    "infairness>=0.2.2",
    "jsonschema>=4.17.0",
    "langchain>=0.2.0",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langgraph>=0.1.0",
    "matplotlib>=3.7.0",
    "neo4j>=5.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pip>=25.1.1",
    "plotly>=5.14.0",
    "psutil>=5.9.0",
    "pytest>=7.3.0",
    "pytest-cov>=4.0.0",
    "pyyaml>=6.0.0",
    "scipy>=1.10.0",
    "spacy>=3.5.0",
    "streamlit>=1.22.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.4.0",
    "transformers>=4.30.0",
]
